requirements.txt
matplotlib==3.2.2
numpy==1.21.6
pandas==1.3.5
py2neo==2021.2.4
pytz==2022.2.1
PyYAML==6.0
scikit_learn==1.1.2
torch==1.11.0
tqdm==4.64.0
utils==1.0.1
fasttext==0.9.3
dgl==2.2.1
